replaced some text with icons in UI and display of player's life
