+++ src/gameData/RoundInitData.ts
@@ -1,9 +1,13 @@
-type RoundInitData = {
+import { Physics } from "phaser";
+import { Player } from "../entities/Player";
+
+export type RoundInitData = {
     round: number;
 }
 
-type UserInterfaceData = {
+export type UserInterfaceData = {
     round: number;
     enemiesLeft: number;
-    enemies: Phaser.Physics.Arcade.Group
+    enemies: Physics.Arcade.Group;
+    player: Player;
 }
